requirements.txt
ebooklib
bs4
opencc
patool
flask
line-bot-sdk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib